+++ js/main.js
@@ -1,5 +1,5 @@
 import { renderThumbnails } from './thumbnail_render.js';
 import './photo_upload_form.js';
-import './validation_description.js';
+import './validation.js';
 
 renderThumbnails();
