+++ js/photo_upload_form.js
@@ -1,6 +1,5 @@
 import { isEscapeKey } from './util.js';
-import { validateHashtags, errorHashtags } from './validation_hashtag.js';
-import { validateDescription, errorDescription } from './validation_description.js';
+import { validateDescription, validateHashtags, errorMessage } from './validation.js';
 
 const body = document.querySelector('body');
 const formUpload = document.querySelector('.img-upload__overlay');
@@ -17,37 +16,28 @@ const pristine = new Pristine(imgUploadForm, {
   errorTextClass: 'img-upload__field-wrapper--error'
 });
 
-function onHashtagsInput () {
+const onHashtagsInput = () => {
   if (pristine.validate()) {
     imgUploadSubmit.disabled = false;
-  } else {
-    imgUploadSubmit.disabled = true;
+    return;
   }
-}
+  imgUploadSubmit.disabled = true;
 
-function onDescriptionInput () {
+};
+
+const onDescriptionInput = () => {
   if (pristine.validate()) {
     imgUploadSubmit.disabled = false;
-  } else {
-    imgUploadSubmit.disabled = true;
-  }
-}
-
-const onDocumentKeydown = (event) => {
-  if (isEscapeKey(event)) {
-    if (document.activeElement === inputHashtags || document.activeElement === inputDescription) {
-      event.stopPrepagation();
-    } else {
-      closeFormUpload();// eslint-disable-line
-    }
-  }
+    return;
+  } imgUploadSubmit.disabled = true;
 };
 
-const onClickCloseBtnForm = () =>{
-  closeFormUpload();// eslint-disable-line
-};
+const onDocumentKeydown = (event) => isEscapeKey(event) && (document.activeElement === inputHashtags || document.activeElement === inputDescription ? event.stopPropagation() : closeFormUpload());// eslint-disable-line
 
-const onClickFormUpload = () =>{
+const onClickCloseBtnForm = () => closeFormUpload();// eslint-disable-line
+
+
+const onClickFormUpload = () => {
   body.classList.add('modal-open');
   formUpload.classList.remove('hidden');
   closeBtnForm.addEventListener('click', onClickCloseBtnForm);
@@ -63,8 +53,8 @@ const closeFormUpload = () => {
 
 openBtnForm.addEventListener('click', onClickFormUpload);
 
-pristine.addValidator(inputHashtags, validateHashtags, errorHashtags, 1, false);
+pristine.addValidator(inputHashtags, validateHashtags, errorMessage, 1, false);
 inputHashtags.addEventListener('input', onHashtagsInput);
 
-pristine.addValidator(inputDescription, validateDescription, errorDescription, 2, false);
+pristine.addValidator(inputDescription, validateDescription, errorMessage, 2, false);
 inputDescription.addEventListener('input', onDescriptionInput);
